requirements.txt
asgiref==3.5.2
async-timeout==4.0.2
attrs==22.1.0
autobahn==22.7.1
Automat==22.10.0
bidict==0.22.0
cachetools==5.2.0
certifi==2022.9.24
cffi==1.15.1
channels==4.0.0
channels-redis==4.0.0
charset-normalizer==2.1.1
constantly==15.1.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==38.0.1
daphne==4.0.0
defusedxml==0.7.1
Deprecated==1.2.13
diff-match-patch==20200713
Django==4.1.2
django-cors-headers==3.13.0
django-import-export==3.0.1
djangochannelsrestframework==1.1.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
drf-yasg==1.21.4
et-xmlfile==1.1.0
google-api-core==2.10.2
google-api-python-client==2.65.0
google-auth==2.14.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.7.1
googleapis-common-protos==1.56.4
gunicorn==20.1.0
httplib2==0.21.0
hyperlink==21.0.0
idna==3.4
incremental==22.10.0
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
MarkupPy==1.14
MarkupSafe==2.1.1
monotonic==1.5
msgpack==1.0.4
oauthlib==3.2.2
odfpy==1.4.1
openpyxl==3.0.10
packaging==21.3
Pillow==9.2.0
protobuf==4.21.9
psycopg2==2.9.4
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
PyJWT==2.6.0
pyOpenSSL==22.1.0
pyparsing==3.0.9
python-engineio==4.3.4
python-socketio==5.7.2
python3-openid==3.2.0
pytz==2022.4
PyYAML==6.0
redis==4.3.4
requests==2.28.1
requests-oauthlib==1.3.1
rsa==4.9
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.6
service-identity==21.1.0
six==1.16.0
sqlparse==0.4.3
tablib==3.2.1
Twisted==22.10.0
twisted-iocpsupport==1.0.2
txaio==22.2.1
types-cryptography==3.3.23.1
typing_extensions==4.4.0
tzdata==2022.5
uritemplate==4.1.1
urllib3==1.26.12
whitenoise==6.2.0
wrapt==1.14.1
xlrd==2.0.1
xlwt==1.3.0
zope.interface==5.5.1
